requirements.txt
SQLAlchemy
SQLAlchemy-Utils
SQLAlchemy-Searchable
Jinja2
joblib
jsonschema
numpy
pandas
PyMySQL
scipy
simplejson
urllib3
kafka-python
dask
distributed
shap
keras
openmldb
redis
jupyterlab
pyarrow
fastparquet
Flask==1.0.3
Flask-RESTful==0.3.9
markupsafe==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Werkzeug==0.16.0
gunicorn==20.1.0
